# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "E-Learning Platform API Documentation",
        "description": "Backend API endpoints for React Web and Flutter Mobile application.",
        "contact": {
            "email": "admin@yourdomain.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Local Development Server"
        },
        {
            "url": "https://dev-api.yourdomain.com",
            "description": "Development Server"
        },
        {
            "url": "https://api.yourdomain.com",
            "description": "Production Server"
        }
    ],
    "paths": {
...
        "/api/v1/courses/{slug}": {
            "get": {
                "tags": [
                    "Course Catalog"
                ],
                "summary": "Get full course details by slug",
                "description": "Returns course details including the section and lesson hierarchy. Paid video URLs are hidden from guests.",
                "operationId": "0c45ecab08c0edb745cbb12ced0b8e90",
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "description": "The course slug (e.g., complete-laravel-react-bootcamp)",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Course details retrieved"
                    },
                    "404": {
                        "description": "Course not found"
                    }
                }
            }
        },
        "/api/v1/learn/courses/{slug}/syllabus": {
            "get": {
                "tags": [
                    "Learning Experience"
                ],
                "summary": "Get full course syllabus with playable videos",
                "description": "Returns the sections and lessons. Only enrolled students will see the actual video_url for paid lessons.",
                "operationId": "e452e2026ef23de381015e9405141894",
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "description": "Course slug",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Syllabus retrieved successfully"
                    },
                    "403": {
                        "description": "Not enrolled in this course"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
        "/api/v1/learn/lessons/{id}/complete": {
            "post": {
                "tags": [
                    "Learning Experience"
                ],
                "summary": "Mark a lesson as completed",
                "description": "Records the lesson as finished and returns the updated total course progress percentage.",
                "operationId": "e881eab91c08faea91134f3d4f8dd228",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Lesson ID",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Progress updated successfully"
                    },
                    "403": {
                        "description": "Not enrolled"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/learn/lessons/{id}/time": {
            "put": {
                "tags": [
                    "Learning Experience"
                ],
                "summary": "Sync video player watch time",
                "description": "The frontend video player should call this silently every 10-15 seconds to save the user's resume point.",
                "operationId": "da9861bcfdf092b02c0809f0a6615ec5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Lesson ID",
                        "required": true
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "watch_seconds"
                                ],
                                "properties": {
                                    "watch_seconds": {
                                        "description": "Current video timestamp in seconds",
                                        "type": "integer",
                                        "example": 145
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Watch time synced successfully"
                    },
                    "403": {
                        "description": "Not enrolled"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Enter your Sanctum token to access secured endpoints",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}